requirements.txt
pymysql=1.1.1
python-dotenv=1.0.0
argparse=1.4.0
logging=0.5.1.2
requests=2.31.0
pyodbc
oracledb==1.3.2